chore: refactor webpack configs
