+++ packages/cli/webpack/src/config/client.ts
@@ -1,7 +1,5 @@
-/* eslint-disable max-lines */
 import path from 'path';
 import {
-  fs,
   isDev,
   isProd,
   CHAIN_ID,
@@ -13,23 +11,28 @@ import {
   isFastRefresh,
 } from '@modern-js/utils';
 import type { IAppContext, NormalizedConfig } from '@modern-js/core';
-import webpack, {
-  DefinePlugin,
-  HotModuleReplacementPlugin,
-  ProvidePlugin,
-} from 'webpack';
+import webpack, { DefinePlugin, HotModuleReplacementPlugin } from 'webpack';
 import { Entrypoint } from '@modern-js/types';
 import { template as lodashTemplate } from '@modern-js/utils/lodash';
-import { WebpackManifestPlugin } from '../../compiled/webpack-manifest-plugin';
 import { InlineChunkHtmlPlugin } from '../plugins/inline-html-chunk-plugin';
-import { AppIconPlugin } from '../plugins/app-icon-plugin';
 import { BottomTemplatePlugin } from '../plugins/bottom-template-plugin';
 import { ICON_EXTENSIONS } from '../utils/constants';
 import { BaseWebpackConfig } from './base';
-import { enableBundleAnalyzer } from './shared';
-import { CORE_JS_ENTRY } from './features/babel-loader';
+import { addCoreJsEntry } from './features/babel';
+import { applyManifestPlugin } from './features/manifest';
+import { applyAppIconPlugin } from './features/app-icon';
+import { applyCopyPlugin } from './features/copy';
+import { applyBundleAnalyzerPlugin } from './features/bundle-analyzer';
+import {
+  applyReactRefreshBabelPlugin,
+  applyReactRefreshPlugin,
+} from './features/react-refresh';
+import {
+  applyNodePolyfillProvidePlugin,
+  applyNodePolyfillResolve,
+} from './features/node-polyfill';
 
-const { USE, RULE, ONE_OF, PLUGIN } = CHAIN_ID;
+const { PLUGIN } = CHAIN_ID;
 
 export class ClientWebpackConfig extends BaseWebpackConfig {
   htmlFilename: (name: string) => string;
@@ -53,33 +56,18 @@ export class ClientWebpackConfig extends BaseWebpackConfig {
 
   entry() {
     super.entry();
-    this.addCoreJsEntry();
+    addCoreJsEntry({
+      chain: this.chain,
+      config: this.options,
+    });
   }
 
   resolve() {
     super.resolve();
 
-    // node polyfill
     if (!this.options.output.disableNodePolyfill) {
-      this.chain.resolve.merge({
-        fallback: this.getNodePolyfill(),
-      });
+      applyNodePolyfillResolve(this.chain);
     }
-  }
-
-  private getNodePolyfill(): Record<string, string | false> {
-    const nodeLibsBrowser = require('node-libs-browser');
-    return Object.keys(nodeLibsBrowser).reduce<Record<string, string | false>>(
-      (previous, name) => {
-        if (nodeLibsBrowser[name]) {
-          previous[name] = nodeLibsBrowser[name];
-        } else {
-          previous[name] = false;
-        }
-        return previous;
-      },
-      {},
-    );
   }
 
   private getCustomPublicEnv() {
@@ -114,20 +102,26 @@ export class ClientWebpackConfig extends BaseWebpackConfig {
     ]);
   }
 
-  private isUsingHMR() {
-    return isDev() && this.options.tools?.devServer?.hot !== false;
-  }
-
   plugins() {
     super.plugins();
 
     this.useDefinePlugin();
-    this.useCopyPlugin();
-    this.useFastRefresh();
 
-    if (this.isUsingHMR()) {
+    const isUsingHMR = isDev() && this.options.tools?.devServer?.hot !== false;
+    if (isUsingHMR) {
       this.chain.plugin(PLUGIN.HMR).use(HotModuleReplacementPlugin);
+
+      if (isFastRefresh()) {
+        applyReactRefreshPlugin(this.chain);
+        applyReactRefreshBabelPlugin(this.chain);
+      }
     }
+
+    applyCopyPlugin({
+      chain: this.chain,
+      config: this.options,
+      appDirectory: this.appDirectory,
+    });
 
     const { packageName } = this.appContext as IAppContext & {
       entrypoints: Entrypoint[];
@@ -231,47 +225,13 @@ export class ClientWebpackConfig extends BaseWebpackConfig {
       .plugin(PLUGIN.BOTTOM_TEMPLATE)
       .use(BottomTemplatePlugin, [HtmlWebpackPlugin]);
 
-    // add app icon
-    const appIcon = findExists(
-      ICON_EXTENSIONS.map(ext =>
-        path.resolve(
-          this.appContext.appDirectory,
-          this.options.source.configDir!,
-          `icon.${ext}`,
-        ),
-      ),
-    );
+    applyAppIconPlugin({
+      chain: this.chain,
+      config: this.options,
+      appContext: this.appContext,
+    });
 
-    if (appIcon) {
-      this.chain
-        .plugin(PLUGIN.APP_ICON)
-        .use(AppIconPlugin, [HtmlWebpackPlugin, appIcon]);
-    }
-
-    this.chain.plugin(PLUGIN.MANIFEST).use(WebpackManifestPlugin, [
-      {
-        fileName: 'asset-manifest.json',
-        publicPath: this.chain.output.get('publicPath'),
-        generate: (seed, files, entries) => {
-          const manifestFiles = files.reduce((manifest, file) => {
-            manifest[file.name] = file.path;
-            return manifest;
-          }, seed);
-
-          const entrypointFiles = Object.keys(entries).reduce<string[]>(
-            (previous, name) =>
-              previous.concat(
-                entries[name].filter(fileName => !fileName.endsWith('.map')),
-              ),
-            [],
-          );
-          return {
-            files: manifestFiles,
-            entrypoints: entrypointFiles,
-          } as any;
-        },
-      },
-    ]);
+    applyManifestPlugin({ chain: this.chain });
 
     if (isProd()) {
       const {
@@ -293,111 +253,14 @@ export class ClientWebpackConfig extends BaseWebpackConfig {
 
     // node polyfill
     if (!this.options.output.disableNodePolyfill) {
-      const nodeLibsBrowser = require('node-libs-browser');
-      this.chain.plugin(PLUGIN.NODE_POLYFILL_PROVIDE).use(ProvidePlugin, [
-        {
-          Buffer: [nodeLibsBrowser.buffer, 'Buffer'],
-          process: [nodeLibsBrowser.process],
-        },
-      ]);
+      applyNodePolyfillProvidePlugin(this.chain);
     }
 
     if (this.options.cliOptions?.analyze) {
-      enableBundleAnalyzer(this.chain, 'report.html');
-    }
-  }
-
-  useCopyPlugin() {
-    const configDir = path.resolve(
-      this.appDirectory,
-      this.options.source.configDir!,
-    );
-
-    const patterns = [...(this.options.output.copy || [])];
-    const uploadDir = path.posix.join(configDir.replace(/\\/g, '/'), 'upload');
-    const publicDir = path.posix.join(configDir.replace(/\\/g, '/'), 'public');
-
-    const minifiedJsRexExp = /\.min\.js/;
-    const info = (file: { sourceFilename: string }) => ({
-      // If the file name ends with `.min.js`, we assume it's a compressed file.
-      // So we don't want copy-webpack-plugin to minify it.
-      // ref: https://github.com/webpack-contrib/copy-webpack-plugin#info
-      minimized: minifiedJsRexExp.test(file.sourceFilename),
-    });
-
-    // add the pattern only if the corresponding directory exists,
-    // otherwise it will cause the webpack recompile.
-    if (fs.existsSync(uploadDir)) {
-      patterns.push({
-        info,
-        from: '**/*',
-        to: 'upload',
-        context: uploadDir,
-        noErrorOnMissing: true,
+      applyBundleAnalyzerPlugin({
+        chain: this.chain,
+        reportFilename: 'report.html',
       });
-    }
-
-    if (fs.existsSync(publicDir)) {
-      patterns.push({
-        info,
-        from: '**/*',
-        to: 'public',
-        context: publicDir,
-        noErrorOnMissing: true,
-        // eslint-disable-next-line node/prefer-global/buffer
-        transform: (content: Buffer, absoluteFrom: string) => {
-          if (!/\.html?$/.test(absoluteFrom)) {
-            return content;
-          }
-
-          return lodashTemplate(content.toString('utf8'))({
-            assetPrefix: removeTailSlash(this.chain.output.get('publicPath')),
-          });
-        },
-      });
-    }
-
-    // options.patterns should be a non-empty array
-    if (patterns.length) {
-      const CopyPlugin = require('../../compiled/copy-webpack-plugin');
-      this.chain.plugin(PLUGIN.COPY).use(CopyPlugin, [{ patterns }]);
-    }
-  }
-
-  useFastRefresh() {
-    if (isFastRefresh() && this.isUsingHMR()) {
-      const ReactFastRefreshPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-
-      this.chain.plugin(PLUGIN.REACT_FAST_REFRESH).use(ReactFastRefreshPlugin, [
-        {
-          overlay: false,
-          exclude: [/node_modules/],
-        },
-      ]);
-
-      this.chain.module
-        .rule(RULE.LOADERS)
-        .oneOf(ONE_OF.JS)
-        .use(USE.BABEL)
-        .tap(options => ({
-          ...options,
-          plugins: [
-            ...(options.plugins || []),
-            require.resolve('react-refresh/babel'),
-          ],
-        }));
-    }
-  }
-
-  // add core-js-entry to every entries
-  addCoreJsEntry() {
-    if (this.options.output.polyfill === 'entry') {
-      const entryPoints = Object.keys(this.chain.entryPoints.entries() || {});
-
-      for (const name of entryPoints) {
-        this.chain.entry(name).prepend(CORE_JS_ENTRY);
-      }
     }
   }
 }
-/* eslint-enable max-lines */
