+++ packages/cli/webpack/src/config/features/svgr.ts
@@ -1,6 +1,6 @@
 import { NormalizedConfig } from '@modern-js/core';
 import { CHAIN_ID } from '@modern-js/utils';
-import type WebpackChain from '@modern-js/utils/webpack-chain';
+import type { WebpackChain } from '@modern-js/utils';
 import { SVG_REGEX } from '../../utils/constants';
 
 export function applySvgrLoader({
