+++ packages/cli/webpack/src/config/features/ts.ts
@@ -1,9 +1,10 @@
-import { applyOptionsChain, CHAIN_ID } from '@modern-js/utils';
+import { resolve } from 'path';
+import { chalk, applyOptionsChain, CHAIN_ID } from '@modern-js/utils';
 import type { IAppContext, NormalizedConfig } from '@modern-js/core';
-import type WebpackChain from '@modern-js/utils/webpack-chain';
+import type { WebpackChain } from '@modern-js/utils';
 import { TS_REGEX } from '../../utils/constants';
 import { getUseBuiltIns } from '../../utils/getBabelOptions';
-import { applyScriptCondition } from './babel-loader';
+import { applyScriptCondition } from './babel';
 
 export function applyTsLoader({
   config,
@@ -83,4 +84,42 @@ export function applyTsLoader({
     .use(CHAIN_ID.USE.TS)
     .loader(require.resolve('ts-loader'))
     .options(tsLoaderOptions);
+}
+
+export function applyTsCheckerPlugin({
+  chain,
+  appDirectory,
+}: {
+  chain: WebpackChain;
+  appDirectory: string;
+}) {
+  const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
+
+  chain.plugin(CHAIN_ID.PLUGIN.TS_CHECKER).use(ForkTsCheckerWebpackPlugin, [
+    {
+      typescript: {
+        // avoid OOM issue
+        memoryLimit: 8192,
+        // use tsconfig of user project
+        configFile: resolve(appDirectory, './tsconfig.json'),
+        // use typescript of user project
+        typescriptPath: require.resolve('typescript'),
+      },
+      // only display error messages
+      logger: {
+        // eslint-disable-next-line @typescript-eslint/no-empty-function
+        log() {},
+        error(message: string) {
+          console.error(chalk.red.bold('TYPE'), message);
+        },
+      },
+      issue: {
+        include: [{ file: '**/src/**/*' }],
+        exclude: [
+          { file: '**/*.(spec|test).ts' },
+          { file: '**/node_modules/**/*' },
+        ],
+      },
+    },
+  ]);
 }
