+++ packages/toolkit/utils/src/compiled.ts
@@ -35,6 +35,7 @@ export type { IOptions as GlobOptions } from '../compiled/glob';
 export type { GlobbyOptions } from '../compiled/globby';
 export type { FSWatcher, WatchOptions } from '../compiled/chokidar';
 export type { ExecaError } from '../compiled/execa';
+export type { default as WebpackChain } from '../compiled/webpack-chain';
 
 /**
  * Lazy import some expensive modules that will slow down startup speed.
